Convert report page components to functional components
