+++ src/containers/registrations-list-total/index.ts
@@ -1,0 +1,1 @@
+export * from "./registrations-list-total";
