+++ src/containers/report-download/report-download.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { FlexGroup } from "../../components/layout/flex";
+import { ReportDownloadLink } from "../report-download-link";
+
+export const ReportDownload = () => {
+
+    return (
+        <FlexGroup
+            direction={"vertical"}
+            style={{ paddingRight: "1em", alignItems: "flex-end" }}
+        >
+            <ReportDownloadLink  />
+        </FlexGroup>
+    );
+};
