+++ src/contexts/store-context/index.ts
@@ -1,1 +1,3 @@
 export * from "./store-context";
+export * from "./store-provider";
+export * from "./use-store";
