+++ src/contexts/store-context/store-context.tsx
@@ -1,9 +1,4 @@
-import React, { createContext } from 'react';
+import { createContext } from 'react';
 import { IRootStore } from "../../stores/root-store";
 
 export const StoreContext = createContext<IRootStore>({} as IRootStore);
-
-
-export const StoreProvider = StoreContext.Provider;
-
-export const useStore = () => React.useContext(StoreContext);
