+++ src/contexts/store-context/use-store.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import { StoreContext } from "./store-context";
+
+export const useStore = () => React.useContext(StoreContext);
