+++ src/stores/view-store/use-view-store.ts
@@ -1,11 +1,7 @@
 import { useStore } from "../../contexts/store-context";
-import { useObserver } from "mobx-react-lite";
 
 export const useViewStore = () => {
     const store = useStore();
 
-    return useObserver(() => ({
-        selection: store.view.selection,
-        toggleSelection: store.view.toggleSelection.bind(store.view)
-    }));
+    return store.view;
 };
